fix render

+++ src/main/java/cn/nulladev/hybridcreatures/entity/LlamaBlaze.java
@@ -45,7 +45,7 @@ public class LlamaBlaze extends Monster {
     public static AttributeSupplier.Builder createAttributes() {
         return Monster.createMonsterAttributes()
                 .add(Attributes.ATTACK_DAMAGE, 6.0D)
-                .add(Attributes.MOVEMENT_SPEED, (double)0.23F)
+                .add(Attributes.MOVEMENT_SPEED, 0.23D)
                 .add(Attributes.FOLLOW_RANGE, 48.0D);
     }
